apps/intelligence/SMART-MALL/requirements.txt
# Smart Mall AI Service Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0

# LLM Providers
openai>=1.10.0

# Vector Database
# pymilvus>=2.3.0  # Milvus
# qdrant-client>=1.7.0  # Qdrant

# Caching
# redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
